WIP #821 Refactoring ExecuteJob
